requirements.txt
django-crontab
gql
celery
django-celery-beat
redis
